move stop_docker() into its own file
